CO-generator/requirements.txt
# Core ML
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
sentencepiece
tokenizers
huggingface-hub

# PDF + Document parsing
pdfminer.six
python-pptx
pytesseract
pdf2image
pillow
pypdf
python-docx
PyPDF2

# UI
streamlit>=1.28.0
pandas>=2.0.0

# Utilities
tqdm
scikit-learn>=1.3.0
numpy>=1.24.0

# Vector database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Knowledge Graph (optional - for Neo4j)
# neo4j>=5.0.0  # Uncomment for Neo4j support

# Metrics & Evaluation
rouge-score  # For ROUGE metrics
nltk  # For NLP utilities


# FastAPI framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP client for health checks
requests>=2.31.0

# Pydantic for data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# System monitoring
psutil>=5.9.0

# Background tasks
aiofiles>=23.2.1
